:recycle: move testy main file inside lib/ along with the rest of the library sources
