+++ lib/script_action.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { Testy } = require('../testy');
+const { Testy } = require('./testy');
 const Configuration = require('./configuration');
 
 const ScriptAction = {
