+++ lib/testy.js
@@ -1,18 +1,17 @@
 'use strict';
 
-const libDir = './lib';
-const TestRunner = require(`${libDir}/test_runner`);
-const { Asserter, FailureGenerator, PendingMarker } = require(`${libDir}/asserter`);
-const ConsoleUI = require(`${libDir}/console_ui`);
-const { allFilesMatching, resolvePathFor, errorDetailOf } = require(`${libDir}/utils`);
-const { I18nMessage } = require(`${libDir}/i18n`);
+const TestRunner = require('./test_runner');
+const { Asserter, FailureGenerator, PendingMarker } = require('./asserter');
+const ConsoleUI = require('./console_ui');
+const { allFilesMatching, resolvePathFor, errorDetailOf } = require('./utils');
+const { I18nMessage } = require('./i18n');
 
 const ui = new ConsoleUI(process, console);
 const testRunner = new TestRunner(ui.testRunnerCallbacks());
 
 /**
- * Object used for writing assertions. Assertions are created with method calls to this object.
- * Please refer to the comment of each assertion for more information ({@link Assertion}).
+ * Object used for writing assertions. [Assertions]{@link Assertion} are created with method calls to this object.
+ * Please refer to the comment of each assertion for more information.
  *
  * @example
  * assert.isFalse(3 > 4)
@@ -45,7 +44,7 @@ const fail = new FailureGenerator(testRunner);
 const pending = new PendingMarker(testRunner);
 
 /**
- * Defines a new test. Each test belongs to a test suite and defines assertions in the body.
+ * Defines a new test. Each test belongs to a [test suite]{@link suite} and defines assertions in the body.
  *
  * For info about assertions, take a look at the {@link assert} object.
  *
@@ -56,8 +55,8 @@ const pending = new PendingMarker(testRunner);
  *   assert.areEqual(3 + 4, 7);
  * });
  *
- * @param {String} name How you would like to call the test. Non-empty string.
- * @param {Function} testBody The test definition, written as a zero-argument function.
+ * @param {!string} name how you would like to call the test. Non-empty string.
+ * @param {!function} testBody the test definition, written as a zero-argument function.
  * @returns {void}
  */
 function test(name, testBody) {
@@ -65,7 +64,7 @@ function test(name, testBody) {
 }
 
 /**
- * Defines a new test suite. Suites are expected to define tests inside it.
+ * Defines a new test suite. Suites are expected to define [tests]{@link test} inside it.
  * There can be more than one suite per file, but it is not possible to nest suites.
  *
  * @example
@@ -75,20 +74,49 @@ function test(name, testBody) {
  *   });
  * });
  *
- * @param {String} name How you would like to call the suite. Non-empty string.
- * @param {Function} suiteBody The suite definition, written as a zero-argument function.
+ * @param {!string} name how you would like to call the suite. Non-empty string.
+ * @param {!function} suiteBody the suite definition, written as a zero-argument function.
  * @returns {void}
  */
 function suite(name, suiteBody) {
   testRunner.registerSuite(name, suiteBody, ui.suiteCallbacks());
 }
 
+/**
+ * Specifies a piece of code that should be executed before each {@link test} in a {@link suite}.
+ * Only one before block is allowed per suite. The most common use of this feature is to initialize objects you want
+ * to reference in each test.
+ *
+ * @example
+ * suite('using a before() initializer', () => {
+ *   let number;
+ *
+ *   before(() => {
+ *     number = 42;
+ *   });
+ *
+ *   test('has the initialized value', () => {
+ *     assert.that(number).isEqualTo(42);
+ *   });
+ * });
+ *
+ * @param {!function} initialization a zero argument function containing the code you want to execute before each test.
+ * @returns {void}
+ */
 function before(initialization) {
   testRunner.registerBefore(initialization);
 }
 
-function after(initialization) {
-  testRunner.registerAfter(initialization);
+/**
+ * Specifies a piece of code that should be executed after each {@link test} in a {@link suite}.
+ * Only one after block is allowed per suite. The most common use of this feature is to ensure you are releasing
+ * resources that are used on each test, like files.
+ *
+ * @param {!function} releaseBlock a zero argument function containing the code you want to execute after each test.
+ * @returns {void}
+ */
+function after(releaseBlock) {
+  testRunner.registerAfter(releaseBlock);
 }
 
 class Testy {
