+++ tests/core/assertions/string_match_assertions_test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { suite, test } = require('../../../testy');
+const { suite, test } = require('../../../lib/testy');
 const { resultOfATestWith } = require('../../support/runner_helpers');
 const { expectSuccess, expectFailureOn } = require('../../support/assertion_helpers');
