+++ tests/examples/basic_assertions_test.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { suite, test, assert } = require('../../testy');
+const { suite, test, assert } = require('../../lib/testy');
 
 const emptyFunction = () => { };
 
@@ -9,55 +9,55 @@ suite('testing testy - basic assertions', () => {
     assert.isTrue(1 === 1);
     assert.that(1 === 1).isTrue();
   });
-  
+
   test('something is false', () => {
     assert.isFalse(1 === 0);
     assert.that(1 === 0).isFalse();
   });
-  
+
   test('objects are equal (using ===)', () => {
     assert.that(42).isEqualTo(40 + 2);
     assert.areEqual(42, 40 + 2);
   });
-  
+
   test('objects are not equal (using !==)', () => {
     assert.that(42).isNotEqualTo(41);
     assert.areNotEqual(42, 41);
   });
-  
+
   test('inclusion in collection', () => assert.that([1, 2, 3]).includes(2));
 
   test('value is included in collection', () => assert.that(2).isIncludedIn([1, 2, 3]));
-  
+
   test('not inclusion in collection', () => assert.that([1, 2, 3]).doesNotInclude(4));
 
   test('value is not included in collection', () => assert.that(4).isNotIncludedIn([1, 2, 3]));
-  
+
   test('inclusion of all elements in collection', () => {
     assert.that([1, 2, 3]).includesExactly(2, 3, 1);
   });
-  
+
   test('error checking', () => assert.that(() => {
     throw 'hey!';
   }).raises('hey!'));
-  
+
   // commented so CI can pass - uncomment to see the failure
   // test('tests can fail as well :)', () => assert.that(() => { throw 'hey!'; }).raises("ho!"));
-  
+
   test('no specific error happened', () =>
     assert.that(emptyFunction).doesNotRaise('hey!'),
   );
-  
+
   test('testing that no specific error happened - even if other error occurs', () =>
     assert.that(() => {
       throw 'ho!';
     }).doesNotRaise('hey!'),
   );
-  
+
   test('testing that no error happens at all', () =>
     assert.that(emptyFunction).doesNotRaiseAnyErrors(),
   );
-  
+
   test('object comparison', () =>
     assert.areEqual({ asd: 2, bnm: [1, 2, 3] }, { asd: 2, bnm: [1, 2, 3] }),
   );
